Added documentation, file handling, and equalization.
As it stands, all job outputs will be routed from the archive of a given job into a newly created folder named Validate GWAS Ouptuts. This is done as Winnow takes an entire folder of input files, so the folder itself will be provided to Winnow alongside the OTE file.