Added a field to the argument parser "-sys" or "--system" that allows the user to specify their desired filesystem. Currently this is defaulting to "data.iplantcollaborative.org".
GWAS switch statement was replaced with a block of if/elif statements. This should fix the issue of only FaST-LMM JSONs being returned to the pipeline regardless of the desired GWAS JSON.

Fixed Winnow running issue - although the app was shared, the execution system was not. This has not been remedied and once said systems are shared this should run in its entirety.