Fixed Ridge input issues, known-truth handling, and started a fix for 400 Client Errors on Agave init.
As Ridge only takes a singular PED file as input, the PED file has been extracted and separated from the MAP for the creation of the Ridge JSON. This should have no impact on other JSONs.

For cases in which the known-truth file is not in the user-defined input directory, a new "-ote" or "--known-truth" flag has been added as a command-line argument allowing a user to directly specify a known-truth file in a different directory.