Bug fixes and code cleanup.
Began reformatting of docstrings and cleanup of old code. Rectified a number of issues in JSON building and began work towards modifying to allow for new/different apps.

Next commit will feature a complete repo restructure so as to make pull requests between Developing-Release and master easier for future releases.

TODO incorporate Merger/equalizer into the pipeline so as to allow Winnow to handle other types of GWAS.